+++ source/Tubeshade.Server/Pages/Libraries/Videos/Edit.cshtml.cs
@@ -1,8 +1,11 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Data;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
+using Npgsql;
+using Tubeshade.Data;
 using Tubeshade.Data.Media;
 using Tubeshade.Server.Configuration.Auth;
 
@@ -10,15 +13,18 @@ namespace Tubeshade.Server.Pages.Libraries.Videos;
 
 public sealed class Edit : LibraryPageBase
 {
+    private readonly NpgsqlConnection _connection;
     private readonly VideoRepository _videoRepository;
     private readonly ChannelRepository _channelRepository;
     private readonly LibraryRepository _libraryRepository;
 
     public Edit(
+        NpgsqlConnection connection,
         VideoRepository videoRepository,
         ChannelRepository channelRepository,
         LibraryRepository libraryRepository)
     {
+        _connection = connection;
         _videoRepository = videoRepository;
         _channelRepository = channelRepository;
         _libraryRepository = libraryRepository;
@@ -39,10 +45,14 @@ public sealed class Edit : LibraryPageBase
     {
         var userId = User.GetUserId();
 
-        Entity = await _videoRepository.GetAsync(VideoId, userId, cancellationToken);
-        Files = await _videoRepository.GetFilesAsync(VideoId, userId, cancellationToken);
+        await using var transaction = await _connection.OpenAndBeginTransaction(IsolationLevel.RepeatableRead, cancellationToken);
 
-        Channel = await _channelRepository.GetAsync(Entity.ChannelId, userId, cancellationToken);
-        Library = await _libraryRepository.GetAsync(LibraryId, userId, cancellationToken);
+        Entity = await _videoRepository.GetAsync(VideoId, userId, transaction);
+        Files = await _videoRepository.GetFilesAsync(VideoId, userId, transaction, cancellationToken);
+
+        Channel = await _channelRepository.GetAsync(Entity.ChannelId, userId, transaction);
+        Library = await _libraryRepository.GetAsync(LibraryId, userId, transaction);
+
+        await transaction.CommitAsync(cancellationToken);
     }
 }