+++ source/Tubeshade.Server/Pages/Shared/StringExtensions.cs
@@ -1,14 +1,12 @@
-﻿using System.Text.Json;
+﻿using System;
+using System.Text.Json;
+using System.Text.RegularExpressions;
 using Tubeshade.Server.V1.Models;
-using static System.StringSplitOptions;
 
 namespace Tubeshade.Server.Pages.Shared;
 
-public static class StringExtensions
+public static partial class StringExtensions
 {
-    private static readonly string[] LineDelimiters = ["\r\n", "\n"];
-    private static readonly string[] ParagraphDelimiters = ["\r\n\r\n", "\n\n"];
-
     extension(string text)
     {
         public string ToJson()
@@ -16,14 +14,40 @@ public static class StringExtensions
             return JsonSerializer.Serialize(text, SerializerContext.Default.String);
         }
 
-        public string[] ToParagraphs()
+        public string? GetFirstParagraph()
         {
-            return text.Split(ParagraphDelimiters, RemoveEmptyEntries | TrimEntries);
+            var span = text.AsSpan();
+
+            foreach (var splitRange in ParagraphSplit().EnumerateSplits(span))
+            {
+                var splitSpan = span[splitRange];
+                if (splitSpan.IsEmpty || splitSpan.IsWhiteSpace())
+                {
+                    continue;
+                }
+
+                return splitSpan.ToString();
+            }
+
+            return null;
         }
 
-        public string[] ToLines()
+        public int GetLineCount()
         {
-            return text.Split(LineDelimiters, None);
+            var count = 0;
+
+            foreach (var _ in LineSplit().EnumerateSplits(text))
+            {
+                count++;
+            }
+
+            return count;
         }
     }
+
+    [GeneratedRegex(@"(?:\r?\n){1}", RegexOptions.None, 100)]
+    public static partial Regex LineSplit();
+
+    [GeneratedRegex(@"\s*(?:\r?\n){2,}\s*", RegexOptions.None, 100)]
+    public static partial Regex ParagraphSplit();
 }